Containerize the cookbook code for TensorRT
Instead of running the cookbook on bare-metal and launching a container
to run the benchmark, the cookbook itself should run in a container and
spawn the benchmark process inside the running container, returning the
results to the working directory on the host. This minimizes the setup
required to run benchmarks as new environments only need to build a
Docker image and start a container.

Signed-Off-By: Robert Clark <robert.d.clark@hpe.com>